Cleanup: Modularize main.js by delegating functions and features
