+++ index.html
@@ -27,6 +27,7 @@
         </div>
 
         <!-- ChartTypeSelecotr & Searchfield / favorites -->
+        <!-- Dropdown with Icons -->
         <div class="row dropdownSearchfield-row">
           <div class="col-2 d-flex align-items justify-content-center">
             <div class="dropdown" id="chartDropdown">
@@ -61,6 +62,7 @@
               </ul>
             </div>
           </div>
+          <!-- searchbar with Fav & Search button -->
           <div class="col-8">
             <div class="SearchAndFavorites">
               <button id="addFavoriteButton" class="btn">
@@ -89,7 +91,6 @@
         </div>
 
         <!-- Slider Controls Group -->
-        <!-- <div class="slider-group-wrapper"> -->
         <div class="row mode-switch-row">
           <!-- Mode Switch -->
           <div
