+++ src/styles/style.css
@@ -232,31 +232,31 @@ canvas#chartCanvas {
     Favorites
 ======================== */
 
-.favorites-list {
+#favoritesList {
   display: flex;
   flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 10rem;
-  margin-bottom: 10px;
   gap: 0.5rem;
+  justify-content: center;
   list-style: none;
-  min-height: 2.5rem;
+  margin: 0;
+  padding-bottom: 1rem;
+  min-height: 52px;
 }
 
-.favorites-list li {
-  padding: 0.5rem 0.75rem;
+#favoritesList li {
   background-color: var(--bs-secondary-bg);
   color: var(--bs-body-color);
-  border-radius: 4px;
+  padding: 0.3rem 0.75rem;
+  border-radius: 0.375rem;
   cursor: pointer;
-  transition: background 0.2s;
+  font-weight: 500;
+  transition: background-color 0.2s;
 }
 
-.favorites-list li:hover {
-  background: var(--bs-tertiary-bg);
+#favoritesList li:hover {
+  background-color: var(--bs-primary);
+  color: white;
 }
-
 /* ========================
     table
 ======================== */
