+++ src/ui/favorites.js
@@ -86,101 +86,3 @@ export function renderFavorites(
     favoritesListEl.appendChild(li);
   });
 }
-
-//==OG code
-
-// function getFavorites() {
-//   return JSON.parse(localStorage.getItem("favorites") || "[]");
-// }
-
-// function updateFavoriteButtonState() {
-//   const ticker = tickerInput.value.trim().toUpperCase();
-//   const favorites = getFavorites();
-//   const icon = addFavoriteButton.querySelector("i");
-
-//   if (favorites.includes(ticker)) {
-//     icon.className = "bi bi-star-fill";
-//     addFavoriteButton.setAttribute("data-mode", "remove");
-//   } else {
-//     icon.className = "bi bi-star";
-//     addFavoriteButton.setAttribute("data-mode", "add");
-//   }
-// }
-
-// addFavoriteButton.addEventListener("click", () => {
-//   requestAnimationFrame(() => {
-//     const ticker = tickerInput.value.trim().toUpperCase();
-//     if (!ticker) return;
-
-//     const mode = addFavoriteButton.getAttribute("data-mode");
-
-//     if (mode === "remove") {
-//       removeFromFavorites(ticker);
-//     } else {
-//       const favorites = getFavorites();
-//       if (!favorites.includes(ticker)) {
-//         favorites.push(ticker);
-//         localStorage.setItem("favorites", JSON.stringify(favorites));
-//         console.log(`Added ${ticker}`);
-//         renderFavorites();
-//       }
-//     }
-
-//     updateFavoriteButtonState();
-//   });
-// });
-
-// function removeFromFavorites(ticker) {
-//   const favorites = getFavorites();
-//   const updated = favorites.filter((item) => item !== ticker);
-//   localStorage.setItem("favorites", JSON.stringify(updated));
-//   updateFavoriteButtonState();
-//   renderFavorites();
-// }
-
-// function renderFavorites() {
-//   const favorites = getFavorites();
-//   const list = document.getElementById("favoritesList");
-//   list.innerHTML = "";
-
-//   favorites.forEach((ticker) => {
-//     const li = document.createElement("li");
-//     li.className =
-//       "list-group-item d-flex justify-content-between align-items-center";
-
-//     const span = document.createElement("span");
-//     span.textContent = ticker;
-//     span.style.cursor = "pointer";
-//     span.addEventListener("click", () => {
-//     list.appendChild(li);
-//   });
-// }
-
-// // Trigger state update when typing in input
-// tickerInput.addEventListener("input", () => {
-//   updateFavoriteButtonState();
-// });
-
-// // Initial render
-// renderFavorites();
-// updateFavoriteButtonState();
-//   });
-// }
-
-// // Trigger state update when typing in input
-// tickerInput.addEventListener("input", () => {
-//   updateFavoriteButtonState();
-// });
-
-// // Initial render
-// renderFavorites();
-// updateFavoriteButtonState();
-
-// // Trigger state update when typing in input
-// tickerInput.addEventListener("input", () => {
-//   updateFavoriteButtonState();
-// });
-
-// // Initial render
-// renderFavorites();
-// updateFavoriteButtonState();
